UmapitaSaveDialogBox を umapita_save_dialog_box.{h,cpp} に分離
